chore: update redux to use redux toolkit, address PR comments
