+++ frontend/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^1.8.0",
     "@testing-library/jest-dom": "5.15.0",
     "@testing-library/react": "11.2.7",
     "@testing-library/user-event": "12.8.3",
@@ -42,11 +43,11 @@
     ]
   },
   "devDependencies": {
-    "@types/react-redux": "7.1.20",
-    "@types/rebass": "4.0.10",
     "@types/jest": "26.0.24",
     "@types/node": "12.20.36",
     "@types/react": "17.0.34",
-    "@types/react-dom": "17.0.11"
+    "@types/react-dom": "17.0.11",
+    "@types/react-redux": "7.1.20",
+    "@types/rebass": "4.0.10"
   }
 }
