+++ frontend/src/redux/store.ts
@@ -1,4 +1,7 @@
-import { createStore } from "redux"
-import { reducer } from "./reducer"
+import { configureStore } from "@reduxjs/toolkit"
+import { reducer } from "./slice"
 
-export const store = createStore(reducer)
+export const store = configureStore({ reducer })
+
+export type RootState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
